Extract custom types to own module, update folder structure
