+++ project/extensions/xtypejs-extension-custom-types/test/.jshintrc
@@ -1,0 +1,11 @@
+{
+    "browser"   : false,
+    "devel"	    : false,
+    "validthis" : true,
+    "eqeqeq"    : true,
+    "eqnull"    : true,
+    "curly"     : true,
+    "boss"      : false,
+    "expr"      : false,
+    "asi"       : false
+}
