+++ project/extensions/xtypejs-extension-custom-types/test/jasmine.json
@@ -1,0 +1,3 @@
+{
+    "spec_dir": "./"
+}
