+++ project/xtypejs/.gitignore
@@ -1,0 +1,9 @@
+/spec/
+/node_modules/
+/bower_components/
+/build/
+/.idea/
+/.project/
+/.settings/
+/.vscode/
+/jsconfig.json
