+++ project/xtypejs/package.json
@@ -1,0 +1,51 @@
+{
+  "name": "xtypejs",
+  "version": "0.5.0",
+  "description": "Elegant, highly efficient data validation for JavaScript Apps.",
+  "main": "dist/xtype",
+  "keywords": [
+    "type",
+    "xtype",
+    "xtypejs",
+    "typeof",
+    "data-type",
+    "validation"
+  ],
+  "author": {
+    "name": "Lucas Ononiwu"
+  },
+  "license": "MIT",
+  "homepage": "http://xtype.js.org",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/lucono/xtypejs.git"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "grunt": "^0.4.5",
+    "grunt-contrib-clean": "^0.6.0",
+    "grunt-contrib-copy": "^0.8.0",
+    "grunt-contrib-jshint": "^0.11.1",
+    "grunt-contrib-uglify": "^0.8.0",
+    "grunt-karma": "^0.10.1",
+    "grunt-shell": "^1.1.2",
+    "grunt-string-replace": "^1.2.0",
+    "jasmine": "^2.2.1",
+    "jasmine-core": "^2.2.0",
+    "jshint": "^2.6.3",
+    "jshint-html-reporter": "^0.2.3",
+    "karma": "^0.12.31",
+    "karma-chrome-launcher": "^0.1.7",
+    "karma-firefox-launcher": "^0.1.4",
+    "karma-coverage": "^0.2.7",
+    "karma-html-reporter": "^0.2.6",
+    "karma-jasmine": "^0.3.5",
+    "karma-spec-reporter": "0.0.16"
+  },
+  "scripts": {
+    "test": "grunt build test"
+  },
+  "bugs": {
+    "url": "https://github.com/lucono/xtypejs/issues"
+  }
+}
