+++ project/xtypejs/test/config/.jshintrc
