+++ project/xtypejs/test/index.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+	<meta charset="utf-8">
+	<title>xtypejs - Specs</title>
+	
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato:100,300,400,700,900">
+    <link rel="stylesheet" href="../../../shared/vendor/jasmine/jasmine.css">
+    <link rel="stylesheet" href="../../../shared/assets/css/xtype.css">
+	
+	<script src="../../../shared/vendor/jasmine/jasmine.js"></script>
+	<script src="../../../shared/vendor/jasmine/jasmine-html.js"></script>
+	<script src="../../../shared/vendor/jasmine/boot.js"></script>
+	
+	<script src="../../../shared/test/test-util.js"></script>
+	<script src="../xtype.js"></script>
+	<script src="spec/xtypejs-spec.js"></script>
+</head>
+
+<body>
+    <div class="top-section">
+        <a href="http://xtype.js.org"><img id="logo" src="../../../shared/assets/img/xtypejs-logo.png"></a>
+        <div class="project-address">
+            <a href="http://xtype.js.org">xtype.js.org</a>
+            <div class="description">Elegant, highly efficient data validation for JavaScript Apps.</div>
+        </div>
+    </div>    
+</body>
+
+</html>
