+++ project/xtypejs/test/spec/xtypejs-spec-node-minified.js
@@ -1,0 +1,8 @@
+
+(function() {
+    var xtypejsTestUtil = require('../../../../shared/test/test-util'),
+        xtypejsMinified = require('../../dist/xtype.min'),
+        specs = require('./xtypejs-spec');
+    
+    specs(xtypejsTestUtil, xtypejsMinified);
+})();
