+++ project/xtypejs/test/spec/xtypejs-spec-node-source.js
@@ -1,0 +1,8 @@
+
+(function() {
+    var xtypejsTestUtil = require('../../../../shared/test/test-util'),
+        xtypejsSource = require('../../dist/xtype'),
+        specs = require('./xtypejs-spec');
+    
+    specs(xtypejsTestUtil, xtypejsSource);
+})();
