--- test/spec/index.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-	<meta charset="utf-8">
-	<title>xtypejs - Specs</title>
-	
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato:100,300,400,700,900">
-    <link rel="stylesheet" href="../vendor/jasmine/jasmine.css">
-    <link rel="stylesheet" href="../assets/xtype.css">
-	
-	<script src="../vendor/jasmine/jasmine.js"></script>
-	<script src="../vendor/jasmine/jasmine-html.js"></script>
-	<script src="../vendor/jasmine/boot.js"></script>
-	
-	<script src="../../xtype.js"></script>	
-	<script src="xtypejs-spec.js"></script>
-</head>
-
-<body>
-    <div class="top-section">
-        <a href="http://xtype.js.org"><img id="logo" src="http://xtype.js.org/assets/img/xtypejs-logo.png"></a>
-        <div class="project-address">
-            <a href="http://xtype.js.org">xtype.js.org</a>
-            <div class="description">Elegant, highly efficient data validation for JavaScript Apps.</div>
-        </div>
-    </div>    
-</body>
-
-</html>
