--- test/spec/typejs-spec-node-minified.js
@@ -1,6 +1,0 @@
-
-(function() {
-    var xtypejsMinified = require('../../dist/xtype.min');
-    var specs = require('./xtypejs-spec');
-    specs(xtypejsMinified);    
-})();
