--- test/spec/typejs-spec-node-source.js
@@ -1,6 +1,0 @@
-
-(function() {
-    var xtypejsSource = require('../../dist/xtype');
-    var specs = require('./xtypejs-spec');
-    specs(xtypejsSource);
-})();
